SVM: Make ret() a special svc, not a syscall
Return can't be a syscall, since it needs to preserve r0 and r1.
This patch swizzles around our encodings a bit such that SVC 0
is now a return (it was previously an always-illegal indirect encoding)
and syscall #0 is now available for other use. I stuck abort() there,
to help keep the code size of ASSERT() macros low.