SVMMemoryLayout bugfix for code symbols across multiple text segments
There were some places where SVMMemoryLayout was using the offset from
the beginning of the segment, when it really should have been using
a shortened form of the actual VA for a symbol.

This patch makes the VA calculation more explicit, and fixes both
symbol address resolution and FNSTACK directives for applications
that have multiple text segments.

The bug occurred for me when compiling with -disable-inlining. The
resulting C++ static initializers by default live in a different
text section.