Another 64-bit bugfix
It is not legal to squash registers r8, r9, pc, sp, or fp, since those
registers must always contain valid physical addresses.

This ties in strongly with the conversation Liam and I had over email.
One solution would be to do a virtual->physical translation on exit
from an exception, but that's awfully heavyweight. We'll probably just
save these (as full reg_t's) somewhere trusted, not in user RAM.