Fix another subtle 64-bit issue which caused the sandwich pathfinding crash
This root-caused to some code which was comparing two stack addresses.
On 64-bit, one of them ended up as a 32-bit VA due to having been squashed
when it was spilled to the stack, whereas the other was still a 64-bit PA.

This is a circumstance that should never occur, and which can't occur on
real hardware or on 32-bit hosts, both of which will always use 32-bit PAs
for stack addresses.

To keep this consistency property on 64-bit simulation without diverging
too much from hardware, this patch squashes addresses during ADDsp. Now
all stack pointers should end up as 32-bit VAs. This is different from
hardware, where they're 32-bit PAs, but 32-bit hosts will still produce
32-bit PAs in simulation, so we still have good code coverage in sim
builds.