Firmware flash block auto-erase
This is a protocol-breaking change to the cube firmware.

The explicit block erase command has been removed from the protocol,
and instead replaced with an auto-erase behavior which is similar
to auto-erase in Stir, except this time it's implemented fully in
the firmware.

During decompression, at every tile boundary, we give the flash_hal
a chance to check whether we've reached a sector boundary. If so,
it does a synchronous erase right there.

Old loadstreams will no longer decompress properly, since although
the erase codes are no longer necessary, the cube firmware now has
no idea how to skip them.