Stir: Leave chromakey on all the time
We may add an explicit option to disable in the future, but this
was causing problems when fully opaque tiles were put on BG1 or in
a sprite. Really, the chromakey option needs to match a tile's use,
not its contents.

But the penalty for leaving chromaKey on all the time is pretty low-
we lose a little bit of our green colorspace, but this is hardly
noticeable.