Fix assetgroup done-ness detection
The bug in multi-asset-group "hello" was triggered when a flash reset
was sent while an earlier asset group was still being written to flash
from the cube's loadstream buffer. Resets are synchronous, and they'll
trash any data that's still in the buffer.

I was about to fix this on the cube side, by making resets more async.
But actually, this problem is really rooted in our detection of asset
group done-ness. We want userspace to only see a group as done if it's
been fully written to flash, not just fully streamed over the radio.

So, this is a master-only change which makes the done-ness detector
also look for a fully drained decoder buffer. Now the demo is working!