Blocking cube HWID accessor
Previously, getHWID was not especially useful on its own. It could
tell you the HWID if we happened to have already gleaned it during
an asset download, but it couldn't request the download on its own.

This was especially noticeable in the sensors demo, when the HWID
started out as all zeroes until the real HWID was retrieved.

Now the accessors, in firmware and in userspace, know how to request
the HWID and block until it's been received.