More progress on SVM LTI Metadata
We still need a key table. I was thinking about generating that in
SVMELFProgramWriter, but really it should be possible to do that in
a separate Transform pass and do it much more cleanly. To keep the
alignment guarantees, that pass will need to know the size of our
ELF header, so that's now public in SVMMemoryLayout.h.