Explicit format string support for _SYS_lti_metadata()
I was hoping to discern all types from the type information
present in the LLVM IR.. unfortunately, in C all arguments
to variadic functions are promoted to int width. So, it wasn't
possible to include 8- or 16-bit values in metadata structs.

So, now this takes a format string akin to python's struct.pack().