More linker magic, to make Metadata::icon() work
I wanted Metadata::icon() to work with a simple
PinnedAssetImage reference, and no special support from
stir or from game authors. I also really want the metadata
to store a VA to the asset group header in flash, not to
the _SYSAssetGroup in RAM, so that there's no dependency
on the contents of user RAM.

This adds an _SYS_lti_initializer() intrinsic which retrieves
the static initializer from a global variable, and returns a
pointer to a new copy of it.