New world order of time APIs
This patch does away with the abuse-prone System::clock(), and replaces
it with a suite of higher-level time classes. This includes a basic
SystemTime class, which is a wrapper around the 64-bit nanosecond clock,
and a TimeDelta class which can efficiently represent time differences
with millisecond precision. It also includes some handy utility classes,
including a version of the Ticker that membrane uses extensively.

These classes have operator overloads such that most of the time you
can just treat TimeDelta like it's a floating point value in seconds.
But when you do arithmetic on floats and TimeDeltas together, we prefer
to convert the floats to TimeDeltas first. This way, in the very common
cases, the floating point math goes away entirely during constant folding.