Firmware header file cleanup
Now that the SDK and firmware are actually totally separate entities,
we can police our use of header files more thoroughly. We shouldn't
be exposing internal details of the firmware to the public SDK, and
we shouldn't be relying on userspace utilities in the firmware. So,
with this change, the firmware is only using sifteo/abi.h, and no other
SDK headers.