New auto-erase scheme for buddy cube flash_decoder
The old auto-erase scheme involved an explicit check,
in the codec, at every tile boundary. This wasn't sufficient, since
it seems that STIR knows how to create RLE codes that cross tile
boundaries, in some cases. In any case, it'd be nice to not depend
on the codec to handle autoerase.

So, this patch moves autoerase entirely into the HAL layer. It
keeps a bit flag indicating whether a full auto-erase check is
necessary. This bit is set when we change addresses manually, or
when lat1 overflows. So, we incur a one-instruction penalty per
word-write. Probably fine.