Use auto-generated symbolic syscall names in Slinky
Previously, I had been hardcoding the syscall numbers for any calls that
slinky needed to generate on its own. This sucked, since it duplicated
information from abi.h, and it makes it very very annoying to reorder the
syscalls at all.

This patch adds a Python script to generate a table at build-time,
so that we can use #define'd names for all syscalls.

Now the *only* place we ever define syscall numbers is in abi.h. Like it should be.