Breakpoint support, by runtime flash patching
This patch implements support for breakpoints. It isn't totally
useful on its own, since GDB still tries to use single-step quite
a lot, and that isn't implemented yet.

To support breakpoints, we give SvmDebugger a way to invalidate
flash pages, and apply patches to them as they load back in from
external flash.