Renumber syscalls, fix event dispatch
Many folks on the game team reported that neighbors stopped working
recently; turns out the problem was a more general issue with event
dispatch, but not one that completely broke events under all cases.

This turns out to be, of all things, a syscall numbering bug.
Normally, calls and syscalls are required to have their return
address bundle-aligned so that it can be stored in a CallFrame
safely. There's an exception made for low syscals in the #0-63
range, since in the original numbering these calls never caused
events, and they were intended to be the most commonly used calls,
so we'd be paying a higher penalty for all those alignment slots.

This change restores the original design by coding the above
requirement into our renumber-syscalls tool. After making this
change, I collected a fresh syscall frequency trace and ran it.
This checkin includes the new abi.h, which will be used by
the runtime and by Slinky.