Maybe fix intermittent win32 build failures
I was using sed and tee as an awful hack to normalize path separators
on Windows. But I think this combination for in-place editing was
racy on Mac OS. On Windows it seemed like 'tee' would always open
the new file after sed had opened the old file, and they wouldn't
conflict, on Mac OS sometimes I'm seeing the new file get truncated.

Perl has more reliable in-place editing, and it should be available
on MSYS and Mac OS. Try that instead.