One-liner fix for "Vec2 bug"
This bug was occurring when an aggregate type was stored on the stack
at a non-64-bit aligned address. In some cases, the optimizer would
use this assumption in order to replace ADD with OR. This patch fixes
the root cause by informing LLVM of our actual platform alignments.
In short, 32-bit for everything.

This following example illustrated the bug:

using namespace Sifteo;

void __attribute__ ((noinline)) test()
{
    volatile unsigned idx = 0;
    Int2 vecs[1];
    vecs[idx].set(5, 6);
    LOG(("(%d, %d)\n", vecs[0].x, vecs[0].y));
}

void main()
{
    volatile unsigned pad = 0;
    test();
}