Fix format string errors in ASSERT()s
We were pasting the assert expression into a _SYS_lti_log() format
string, which would break if the expression happened to contain any
"%" characters. This now uses the new value inlining support in
slinky to keep these values out of the format string. As a bonus,
the resulting code is way more intuitive :)