Last minute fix for syscall struct returns
This is an ABI issue that turned up when we switched clang targets:
Clang is now coding struct returns as such, which has LLVM pass a pointer
into the function rather than having the struct value passed out. This
is the case even for small structs that fit in one or two machine
words.

This behavior is actually fine in most cases, but it breaks our
syscall ABI. So, in the interest of keeping syscalls simple and
decoupled from such shenanigans, I changed the last couple of
stragglers to use pure uint32/uint64 return types again.

Because games are still using raw syscalls, they're broken
by this change. But I'm going to break the world again very soon,
so...