Fix a subtle PaintControl bug
If _SYS_finish() is the first thing we do on a newly attached
video buffer, we might take a really long time to get back in
sync because makeSynchronous() sets SYNC_ACK when this isn't
actually valid due to the vbuf's flags not having ever been
flushed to the hardware.

Fixed this by adding an additional flag to keep track of whether
we know the vbuf's rendering flags have been flushed.