Fix a cube deadline scheduling race condition
This happened if the cube thread made enough progress before the
MC thread starts, such that it's already past the first event deadline.
That can be fixed easily by initializing the first deadline to 0, but
then it reveals a secondary problem in which the cubes may stop earlier
than the intended deadline. This is by design; so document it in
beginEvent(), and give the radio code a loop to scoot forward until
the actual deadline if the cubes have stopped early.