Fix a nasty bug in CubeCodec::deltaSample
This was due to an order-of-operations error in the inverse TILE77
transform, which ended up creating false deltas for tiles which only
differed by 0x80 in either byte.

To avoid getting this transform wrong in the future, I added it to
abi.h as a counterpart to _SYS_TILE77, and added some ASSERTs in
deltaSample() to make sure it keeps working.