Concurrent Finish implementation across all cubes
Finish is more complicated than it used to be and, like Paint,
we need a way to run it concurrently across all cubes. Some of
the really big delays Eric has been seeing are due to some
(smaller, but still large-ish) delays per-cube that stack when
more than one cube is waiting.

Because Finish may involve many separate state transitions,
I decided to implement it using a polling approach where the
overall wait loop is in CubeSlots, and the PaintControl exports
a begin() and a poll() method.

With this change, chroma's stalls are down to the expected
level of 1/fpMin, which is the unavoidable threshold for
transitioning from continuous to synchronous rendering. Next
I'll be decreasing this threshold, but that can be a separate patch.