CubeCodec efficiency: Lock exemption ranges
We use locking to keep track of which words have a
potentially-different value than the one we intend to
send, or the one the cube actually has. This avoids
using that word as a sample for a delta code.

BUT, regardless of the lock state, we can always
make delta codes based on words that have been encoded
earlier within the same packet. This is because no userspace
code may run, and no other VRAM modifications may occur
during packet encode. We can take advantage of this to speed
up a common case: Keep track of the last run of consecutive
codes we've generated, and allow any source within that last
run to be used even if it's locked.

This causes Stimulant's demo to use far less radio traffic!