PaintControl: Much faster algorithm for exiting continuous mode
Previously the process of exiting continuous mode was counted
as a synchronization-breaking event, which necessitated the 100ms
fpsLow delay, in order to wait for the cube to render.

Now that we have the FRAME_ACK_CONTINUOUS feedback bit, we can
shift this around such that the process of *entering* continuous
mode is the sync-breaker. If we're sure we've fully entered
continuous mode, we can exit it with low latency.

This should greatly reduce any visible stutters in apps like
Chroma which provoke a lot of sync/async switches.