New CPU flags for emulator
This should fix the deadlock we were seeing with autobuilds...

The build machine has an AMD CPU, so -march=native there was
generating AMD-specific code. Specifically, it was using the LZCNT
instruction which was flat-out returning incorrect results on
Intel. This broke the common CLZ loops in our firmware, turning
them into infinite loops. And thus the firmware hangs and the simulator
makes no forward progress.

Instead, let's use only P4-and-later instructions, with tuning for Core2.