Fix dynamic cube add/remove
The gross thread start/stop hack in System was ridiculously
annoying to reconcile with the new strong assumptions used
by the DeadlineScheduler... so instead of hacking the hack,
I removed it in favor of a fast per-tick-batch mutex, the
Big Cube Lock. While holding the Big Cube Lock, you're
guaranteed that all virtual cube hardware is stopped.

In our case, we use this state to set opt_numCubes or to
initialize cube hardware. Though in the future this would
also be hella useful from inside Lua scripts, which currently
have no way to actually synchronize with cube execution.