Uncached flash access for ELF debug info
The many flash blocks with no valid VA tended to be addresses
outside the mapped ROData segment, typically in the ELF debug
sections. This was getting paged in any time we looked up debug
symbols, or resolved the name of a function.

This change adds a cache-bypassing copy primitive for such things,
and it uses it in ELFDebugInfo.