Make firmware builds mandatory
Now that we have self-contained binary releases of the SDK, there's
less reason to make the full TC build as self-contained as possible.
To reduce fragmentation in our environment, this change makes the
firmware builds mandatory, as well as making Python mandatory for
the build.

This also means that we can build the auto-generated C++ source
for the emulator as part of the build, rather than keeping it
checked in. No more SBT firmware update checkins!