Last minute abi.h refactoring
This started as an intent to fix the cube firmware build on SDCC 3.1.0.
We never needed 64-bit type support in the cube firmware, but other
defs in abi.h need 64-bit types. I thought I'd fixed this with a typedef
in the cube firmware, but SDCC 3.1 added some more aggressive warnings
that now bite us.

So... split up abi.h, and include only the necessary VRAM defs in
the cube firmware.