Use -march=i686 on Windows
Previously we were using pentium4, which seemed to be the right level
of CPU compatibility for our needs. However, Eric found what turned out
to be a pretty nasty compiler bug which causes SSE2 instructions to
operate on data which may not be sufficiently aligned. This was causing
Siftulator to crash on WinXP machines.

Since we don't especially need SSE or SSE2 in Siftulator, it shouldn't
hurt us much to drop back down to i686 there.