New volume block ordering guarantee
The header block is now always guaranteed to come first,
which fixes a serious correctness and security flaw in
the way we iterate over flash volumes.

To implement the fix, i added some nice infrastructure
for borrowing cache memory, and for preparing writes to
pages that haven't yet been assigned an address.