Reference-accurate handling of kNoteOff
kNoteOff doesn't kill a note, it works more like MIDI where the note's
playback future is determined by other factors. In XM a Note Off is
handled by activating the fadeout (if applicable) and disabling looping
(if applicable). The sample continues to play until the sample runs out
(fadout/volume set to 0 technically should still advance sample
playback position, but a future optimization could be to not bother
actually fetching the samples when the volume is 0).