Latency optimization for _SYS_VBF_TRIGGER_ON_FLUSH
This decreases overall paint latency by approximately
one radio-packet-period, which can help a lot in extremely
tight paintUnlimited/finish loops.

The mechanism is simple: Previously we would end up sending
the actual trigger on a future packet, even if there was space
in the TX buffer to include the trigger in the same packet
that caused us to notice that we're done flushing. Now we
try to reenter encodeVRAM() once after vramFlushed() causes
a trigger.