Floating point library unwrapping
Until now, we were effectively wrapping the software floating
point library in a big layer of C++ padding, in the form of
syscall_float.cpp. This was pure C++ code which was written to
be equivalent to the corresponding EABI/libgcc floating point
functions. On emulator builds it converted to native CPU float
support, and on stm32 builds it ended up generating no-op
wrappers around the corresponding soft-float functions.

However, GCC totally sucked at optimizing those wrappers.
Absolutely everything had an unnecessary load/store on it,
plus the wrappers themselves took up code space and stack
space.

This patch eliminates the wrappers. The pure C++ implementations
of the float syscalls have been moved to the emulator source tree,
and I added an aliasing system to the syscall table generator
so we can pipe those syscalls directly where they need to go.

I tried doing the aliasing with our linker script, but that caused
more problems than it solved, since the linker script doesn't seem
to be able to reference library symbols that would otherwise be
garbage collected.

WARNING: Haven't tested this on hardware yet, but the binary
looks right. I should be able to test it on hardware tomorrow.
My hope is that this improves Mandelbrot performance even more :)