Remove ASSERT in Volume::init
It isn't okay to assert that init() only happens once. When we simulate
a restart of the master firmware, we don't have a good way to simulate
resetting the global memory state that would be in the firmware's BSS
and RWDATA segments. Maybe there's a way we could force the firmware's
data to go into separate segments which we could manually re-init at
runtime, but so far the simpler approach has been to just make the
firmware gracefully handle repeated inits.