Stir: Add an additional local palette reduction pass
Before this patch, the only palette reduction we would perform
was global, and we focused on reducing the total colorspace as
much as possible while staying under the MSE limits.

This was good at keeping the global palette consistent throughout
the entire game, but we'd still end up with individual tiles
containing lots of colors. Some of this was mitigated by the
encoder's RLE optimizations, but it was still common to end up
with enormous piles of CM_TRUE tiles.

This new pass focuses *only* on tiles that would otherwise have
too many colors to use a LUT, and it performs strictly local
palette optimization within that individual tile. We use a higher
than usual MSE limit to check the resulting tile.

Tested on chroma. There, we're eliminating 94% of the CM_TRUE tiles.