Fix race condition in simulated TestJig
The write() primitive was not atomic at the packet level, only at the byte level. The most common failure mode was that the first byte of an outgoing packet would get "eaten" by the one spurious I2C read after NAK, so that byte would be permanently ignored by the cube firmware, resulting in a test failure.

This should fix the intermittent test failures in Jenkins.