Fix windows build?
Okay, Maureen hung again in the same spot, even with the latest
code, and even after unwedging it once manually.

The audio tests work fine on my Windows 7 box at home, so I'm not sure what's up with maureen. For now, fix the CI builds by adding the audio tests to a growing list of everyone-but-windows tests.