New unaligned-data-friendly SvmMemory CRC primitive
This patch adds SvmMemory::crcROData(), a new CRC primitive
which uses the hardware Crc32 engine, but includes support for
unaligned memory, and padding the result to a 32-bit (or larger)
alignment.

It turns out that I need all of those features in _SYS_fs_objectWrite,
but it's really a pretty general-purpose thing, and these are capabilities
that would be really helpful in the userspace crc32() function.

So, this replaces crc32, changing its signature to accept any
byte pointer instead of a word pointer.

This patch updates the unit tests, and adds a new reference CRC
implementation for the unit test and the docs.