Implement our own system time functions, avoid focus-stealing
On Mac OS, glfwInit() would end up initializing Cocoa to the point
where the window manager saw us as a GUI process and we'd steal the
focus. This happened even in headless mode, which was super annoying.

This patch adds our own timer primitives so we can finally stop
using glfw for anything other than the actual GUI.