Decrease number of cube slots, from 32 to 24
We've been at 32 cubes for so long, since the only real practical limit for most of the firmware was the width of our bitmaps. I finally found the second technical constraint.

After mulling over the asset slot persistence format for most of the day, I think the best option is to have one LFS object representing an overall allocation map for cube slots, then one LFS object per (cubeID, slot) tuple which can track the actual asset groups in that slot and where they're loaded.

The problem: We have 8 AssetSlots per cube in the current design, and 32 cube slots. That's exactly 256, meaning it would consume every single key in an LFS volume. I'd like this data to all share a single 'system' LFS volume, along with the cube pairing data and whatever else wants to live there.

So, the best solution seemed to be decreasing the number of cube slots down to a more realistic number, so that we can reclaim some of those LFS keys. This gives us 32 spare keys after the asset slot allocator is through, which should be plenty for the rest of the system :)