Make mixer sample rate a compile-time constant
We've been building up more and more assumptions in the system about sample rate, such that it'd be really impractical (and really unnecessary) to change it at runtime. So, let's be honest about that, and make it a compile-time constant. This gives us some benefit in terms of optimization as well as STATIC_ASSERT's to keep track of where these assumptions are being made.