Syscall pointer alignment checks
Another easy but mundane thing that's been sitting on my todo list for a while... syscalls need to check the alignment of any pointers they're passed, and fault if it isn't correct.

This isn't strictly necessary on our current hardware, since the STM32 does support unaligned memory accesses. However it's best not to rely on it, since it's slow and may introduce unwanted complexity in atomic operations. Any correct usermode program shouldn't be passing misaligned pointers to the system anyway, so let's just catch them early.