New I2C state machine layout, with A21 support
This is the first step in adding A21 support to the firmware. I needed to squeeze Yet More Things into the I2C state machine, so I restructured it a little in order to make it more space-efficient, to keep it within the 256-byte limit of an 8-bit state machine.

The actual A21 support in the ISR is implemented as a pair of bits indicating the desired A21 state and the last set A21 state.