Disable checkQuiescentVRAM by default
Having this on unconditionally has outlived its usefulness. Keep the implementation of checkQuiescentVRAM itself, as it's a useful tool for debugging specific problems- but remove the heuristic which called it automatically. The heuristic isn't perfect, and all of the last half dozen or so failures I've debugged have been problems with the check's heuristic rather than representing actual codec problems. We haven't seen any actual codec bugs in quite some time now.