Minor consolidation of Tasks::idle()
I really want to keep the entire idling logic in one place as much as possible, so that callers which wait on a condition can do so simply with:

 while (!condition) Tasks::idle();

I also wanted to try implementing my original thought on how idle() should work, in which we do only one of work() or waitForInterrupt(). I've updated the comments to indicate that this is the intended design, but unfortunately it doesn't work yet because we have no way yet to filter out tasks which are always pending without necessarily making forward progress, like AudioPull.

So, until then, the implementation of idle() always calls both work() and waitForInterrupt(). This patch shouldn't change runtime behavior at all.