ADPCM decoding optimization
This patch modifies the ADPCM codec to be a lot more efficient on ARM:
  - Use multiplication (fast!) instead of a ton of branching
  - It's table driven, as much as is practical
  - Saturate using the SSAT instruction
  - Inline each nybble's decoding

This required changing the rounding rules slightly, meaning we aren't standard IMA ADPCM. But as far as I can tell, this codec was never actually fully compatible with IMA ADPCM anyway, due to the different order in which we update the delta index. So, we'll just stop calling it IMA ADPCM.

Also use the SSAT instruction in the mixer, while we're at it.