Fix buffer overrun in SysLFS::AssetSlotRecord::load
This was showing up as a crash in the assetslot unit test on 32-bit hosts. (The bug was masked on 64-bit)

Due to LFS padding, the FlashLFSObjectIter may report a size greater than sizeof(AssetSlotRecord). The new code to support variable-sized AssetSlotRecords has a flaw in which it always reads the stored size, even if it's larger than the record struct. This is always a security flaw, and an exploit would be trivial with a way to write to SysLFS. Additionally, this bug would crash all on its own if the LFS record size was padded to a size larger than the buffer.