Okay... increase still more
I diff'ed the images again, and this is still coming from just a couple of the most intensely illuminated pixels in the hilight on Ball2. My guess is it's an -ffast-math side-effect in the CIELab code in stir. Not terribly concerned, except insofar as this is breaking my unit tests!