Cube firmware size optimization
I thought we had 200+ bytes free... turns out, though, that sizeprof wasn't accounting for space used by SDCC's standard library. The only useful purpose this library code currently serves was to initialize the stack and clear IRAM. We didn't even use this code in SBT, because it was quicker to do the setup in C code than it was to translate the SDCC library code.

This patch banishes the SDCC standard library entirely. Due to an annoying lack of flexibility in SDCC's handling of main(), this meant going all the way to the metal and rolling our own IVT. On the bright side, there's a lot of empty space in the IVT and we can interleave the in-use portions with our initialization.

Back up to ~230 bytes free now, and no hidden library code to speak of.