Increase erase count ratio threshold
The "ideal" erase count calculations in the unit test don't account for LFS overhead, so we're getting close to the arbitrary limit of 1.5. Increase the limit to 2.0. Any real wear levelling failures are going to cause this ratio to get kind of astronomical anyway, so this still seems quite safe.