Fix flash read-after-write on hardware
Whew.. this seems like it was the underlying cause of a *lot* of system instability as soon as we started writing to flash at runtime. There was no code in place to wait for the write to complete before reading, so we'd end up nondeterministically reading back all 0xFFs from the hardware. This refactors our busyness handling a bit, and fixes the bug