Fix exec() tail-syscall from main() bug
This is a somewhat esoteric bug in the SVM runtime that I hit
while debugging the new launcher...

If you exec() another binary from a tail-syscall, this would cause
SvmRuntime::ret() to immediately return from the new binary,
restarting the launcher.

A similar problem could result from a fault in any tail syscall.

This patch changes the handling of ret() slightly so that we can
cancel the exit operation if we've exec'ed during the syscall,
either due to an expicit exec or due to a fault.