Firmware support for ADPCM initial conditions
This patch modifies the ADPCM compression format to include a
3-byte header with the initial conditions for the decoder. A
forthcoming patch for Stir will generate this header.

This addresses a problems with quality of ADPCM encoding
short samples. With the default initial conditions, ADPCM
could take much too long to adapt, causing severe distortion
in the small samples used by XM modules.

This is intended to allow Stir to search for the best possible
initial conditions for each encoded sample.