Slinky support for compressed data segments
Whew! This is a bit wacky compared to what I'm sure the LLVM authors had in mind... but it works.

After fixups but just before generating the symbol table and debug sections, we do a pass over the program's sections in order to flatten all initialized data into a single buffer. That buffer is compressed, then we create a new section for the compressed data. The original plaintext sections are only included in debug builds, for gdb's use.