fwdeploy: Salt the plaintext binary before encryption
This fixes a potential crypto weakness I noticed: broadly speaking, I'm worried that we may be vulnerable to a class of differential cryptanalysis attacks which involve known (or guessed) differences in two plaintexts. For example, if we released two slightly different firmwares (say, with a tiny bugfix in them) it may be easy for an attacker to guess what change we made in the plaintext. This could then be used to compare the two corresponding cyphertexts, and extract the key.

A much more obvious attack: Since the cyphertext only depends on preceeding bits, not succeeding bits, if you have a pair of firmware images with some small difference in them, you can compare the encrypted images in order to determine where the first difference is in the corresponding plaintexts. So, if we patched a vulnerability in some firmware function for example, it would be downright trivial to determine the physical address of this function!

My fix is pretty simple: I apply a salt near the beginning of the binary, generated using a cryptographically secure RNG. To avoid necessitating a bootloader update, I wedged the salt into a reserved portion of the IVT so that it's near the beginning of the binary, but in otherwise-unused space.

Tested on Windows and Mac. Verified that two consecutive runs of fwdeploy on the same binary generate almost completely different outputs. (Only the first 0x24 bytes match, which is mostly header)