Increase flash decoder FIFO size
Make the small change necessary to allow non-power-of-two sizes,
and bump it up to 80 bytes. Right now, there's a serious producer/consumer
bottleneck. The new decoder is consuming bytes faster than the old one,
and in a somewhat more bursty way. This is reducing our headroom for
tolerating a buffer that's running low.

This change helps a lot, but it looks like we really need the buffer
to be a little bit larger still. I'm going to try some new RAM usage
accounting methods, since RAM is getting tight enough that I'd like
a more clear idea of where it's all going.