Add a static analyzer for cube firmware RAM usage
This computes the worst-case stack sizes
using static analysis. This also introduces
specially formatted comments with instructions
to the static analyzer. Currently they're only
used by sizeprof, but we may be able to use
them to make siftulator much more efficient
in the future.

Right now it shows we're overbudget by 15 bytes!
It's unlikely, but we could get a stack overflow
in the very worst case.