Give up some ground on the load FIFO
This gets us back in budget. Now I can add sizeprof to the default
build, so we stay in check from now on.

I'd still like to make the load buffer larger... but need to figure
out how to do that. I'm toying with the idea of repurposing part of
VRAM temporarily.