Possible siftulator performance fix
This patch might fix the issues with postMessage() that Winston's been seeing. It's hard to tell precisely because my Windows VM gets just so darn slow when siftulator is running, but I think the problem he's seeing is related to lock contention on the Frontend's big instance lock.

This is kind of a bad locking situation on any platform, but my guess is it's much more apparent on Windows because we have coarser-grained sleep there.

This patch reorganizes Frontend::runFrame() so that we can drop the lock before blocking on either OpenGL or sleep().