Explicit ACK retransmit request (Protocol break!)
Whew.. I don't know that I'm totally to the bottom of this set of firmware loading issues, but this appears to be the big one that's getting asset loads stuck frequently. This fixes an old "XXX" comment that I carefully left then forgot about. It's always been possible for asset loading to deadlock due to dropped ACK packets; this is just the first time the conditions were right such that we'd see it super-reliably.