Disable magic HX8353 init
I'm trying to track down this visual interlacing/flickering that's been bugging me with the current TC firmware, and I think it might be a badly balanced gamma table in the initialization code. The LCDs seem to work better if I just leave everything at factory defaults, so I'm trying that for now. The bulk of the initialization sequence is now disabled by default.