Some hopeful changes that may fix our random hangs
I'm having a hell of a time isolating this hang- it's highly
unpredictable, and it seems like the closer I get to the root
cause, the more difficult it is to debug the system without
changing its behavior.

At present, my best guess is that we're triggering some kind
of hardware or software bug related to DMA or SPI, which
causes a DMA from flash memory to never complete.

I *do* know that:

 1. We've only ever seen this crash during audio mixing
 2. The CPU is spending lots of time in waitForDma() while
    we're in this hung state.
 3. When hung, we're no longer calling mapROData() in
    the ADPCM decompressor.

This plus my sampling profile results (via the UART) point
very strongly to a DMA that never completes. I tried to catch
it in the act, using counters that increment when a DMA
starts and complementary counters which increment when
the completion IRQ comes back. But nada. When I put any
form of instrumentation that close to the DMA code, I can't
repro the hang any longer... which makes me even more
suspisious of the timing of these SPI and DMA operations.