Fix a buglet in the radio ISR's clearing of ack_bits
When this code was first written, I don't think we had any higher
priority ISRs, so I wasn't thinking about this code being preemptible.
But now we do, and the RF ISR needs to clear ack_bits earlier in order
to avoid a race condition that may mean dropping ACKs.