Change the flash-reset ack mechanism
While we're breaking protocol anyway, this is something that's been bugging me for a while. The previous acknowledgment mechanism was ambiguous if there's any possibility the FIFO was non-empty at the time of reset. This is usually not an issue, but it seems like it could make it more difficult to recover from errors under some scenarios. We have plenty of spare bits in the packet, so use one of them as a reset ack toggle.