Fix color disaster in tilerom
The tilerom atlas was being generated in 24-bit color, instead of
using the quantized colors we actually store in the ROM. For foreground
colors this isn't a huge difference, but for background colors it's
huge. I had forgotten, but the BG0_ROM implementation uses 8-bit
background colors for various optimization reasons.

This patch fixes the atlas to use device-accurate colors,
modifies the reference images accordingly, and changes some palettes
which were super-ugly in 8-bit color.