Plan to not have a good sincosf() in the forseeable future
I originally wrote the syscalls this way because I was assuming we'd have
a fast algorithm to calculate sin() and cos() simultaneously. Well, these
algorithms exist, but our libm certainly doesn't have one and I can't find
a good free implementation to use.

So, let's just change plans. Keep the sincos syscall for a rainy day,
but implement it directly in terms of sin() and cos(), and add direct
syscalls just for sin and cos.