Big Tasks cleanup
The task array in RAM was kind of wasteful... only a single task (audiopull) ever had a parameter, and its value was guaranteed to be known at compile-time. So, this does away with task paramters entirely. Individual tasks can store things in RAM better than our centralized code could, if they really need to.

I made the audio buffer static, and it's now owned by the mixer.

Additionally, task functions are now dispatched by a switch() instead of a lookup table. We have few enough and sparse enough tasks that this may be a win, and it gives gcc the opportunity to use special jumptable instructions. But most of all, this makes configuration way less error-prone.